Restructure file and formatting
